+++ mobileapp/core/lib/usecases/routedb.dart
@@ -5,9 +5,11 @@ library;
 
 import 'dart:async';
 
+import 'package:crosscuttings/errors.dart';
 import 'package:crosscuttings/logging/logger.dart';
 
 import '../boundaries/ota.dart';
+import '../boundaries/positioning.dart';
 import '../boundaries/presentation.dart';
 import '../boundaries/storage/preferences.dart';
 import '../boundaries/storage/routedb.dart';
@@ -37,6 +39,9 @@ class RouteDbUseCases {
   /// Interface to the download boundary, used for fetching database updates.
   final RouteDbDownloadBoundary _downloadBoundary;
 
+  // interface to the positioning component, used for getting the current location.
+  final PositioningBoundary _positioningBoundary;
+
   /// Interface to the storage boundary component, used for reading and writing application config
   /// settings.
   final AppPreferencesBoundary _preferencesBoundary;
@@ -58,6 +63,7 @@ class RouteDbUseCases {
     required this._downloadBoundary,
     required this._preferencesBoundary,
     required this._systemEnvBoundary,
+    required this._positioningBoundary,
   });
 
   /// Use case: Download the most current route database via OTA and install (import) it.
@@ -144,6 +150,38 @@ class RouteDbUseCases {
       filterByAreaId,
     );
     _presentationBoundary.updateSummitList(summitList);
+  }
+
+  /// Use Case: Switch to the main nearby summits list.
+  Future<void> showNearbySummitsPage() async {
+    _logger.info('Running use case showNearbySummitsPage()');
+    // Display the nearby summits page first
+    _presentationBoundary.showNearbySummits();
+
+    // Try to get the current position
+    GeoPosition currentPosition;
+    try {
+      try {
+        currentPosition = await _positioningBoundary.getCurrentPosition();
+      } on MissingPermission {
+        _logger.debug('Location permission is missing but will be requested');
+        await _positioningBoundary.requestPermissions();
+        currentPosition = await _positioningBoundary.getCurrentPosition();
+      }
+    } on Exception catch (error) {
+      _logger.error('Location not available due to', error);
+      _presentationBoundary.nearbySummitsLocationError(error);
+      return;
+    }
+    _logger.debug('Got location: ', currentPosition);
+
+    // Got a position, now find and display all summits in the vicinity
+    List<(Summit, double)> nearbySummits = await _querySortedSummitList(
+      currentPosition,
+      _maxNearbyDistance,
+      NearbySummitsSortMode.distance,
+    );
+    _presentationBoundary.updateNearbySummits(nearbySummits);
   }
 
   /// Use Case: Show detailed information about the selected summit.