+++ mobileapp/infrastructure_flutter/lib/ui.dart
@@ -90,6 +90,25 @@ class ApplicationUI implements ApplicationUiBoundary {
   }
 
   @override
+  void showNearbySummits(NearbySummitsPageLabels model) {
+    _logger.debug('Displaying nearby summits page');
+    _uiState.resetNotifiers();
+    _uiState.nearbySummitListNotifier.fallbackMessage = model.noDataMessage;
+    _switchToRoute(UiRoute.nearbysummits.toRouteString(), isRoot: true, routeArguments: model);
+  }
+
+  @override
+  void showNearbySummitsError(String message) {
+    _uiState.nearbySummitListNotifier.fallbackMessage = message;
+    _uiState.nearbySummitListNotifier.replaceData(<ListViewItem>[], <ListViewItem>[]);
+  }
+
+  @override
+  void updateNearbySummits(List<ListViewItem> summitItems) {
+    _uiState.nearbySummitListNotifier.replaceData(summitItems, <ListViewItem>[]);
+  }
+
+  @override
   void showSummitDetails(SummitDetailsModel model) {
     _logger.debug('Displaying summit details page');
     _switchToRoute(UiRoute.summitdetails.toRouteString(), isRoot: false, routeArguments: model);